requirements.txt
PyYAML~=6.0
pillow~=9.1.0
kagaconf~=0.1.3
hikari-lightbulb~=2.2.1
hikari~=2.0.0.dev108